+++ library/src/main/java/io/fastpix/media3/info/FastPixPlayerLibraryInfo.kt
@@ -2,5 +2,5 @@ package io.fastpix.media3.info
 
 object FastPixPlayerLibraryInfo {
     const val PLAYER_NAME = "fastpix-android-player"
-    const val PLAYER_VERSION = "2.0.1"
+    const val PLAYER_VERSION = "2.1.0"
 }